+++ src/Driver/main.py
@@ -7,10 +7,27 @@ from PIL import Image
 import threading
 
 
+def get_version():
+    try:
+        string_with_version_in_it = ''
+        with open('CHANGELOG.md','r') as file:
+            for line in file:
+                if '##' in line:
+                    string_with_version_in_it = line
+                    file.close()
+                    break    
+                
+        version = string_with_version_in_it[4:12]
+    except FileNotFoundError:
+        print('CHANGELOG.md Not Found, version will be ""')
+        version = ''
+
+    return version
+
 script_path = os.path.dirname(__file__)
 
 #it is the folder the script is in, will need to change if parent folder is different
-version = __file__.split('\\')[-2]
+#version = __file__.split('\\')[-2]
 
 # Gets the path of icon image
 icon_path = fr"{script_path}\pythonIcon.ico"
@@ -18,7 +35,7 @@ image = Image.open(icon_path)    # Opens the Icon
 #https://stackoverflow.com/questions/6893968/how-to-get-the-return-value-from-a-thread-in-python
 
 def sysIcon():
-    systray = pystray.Icon(name="Python Macro", icon=image, title=f"Python Macro{version}", menu=pystray.Menu(
+    systray = pystray.Icon(name="Python Macro", icon=image, title=f"Python Macro {get_version()}", menu=pystray.Menu(
         pystray.MenuItem("Quit", lambda: macro_driver.stop())
     ))
     systray.run()
@@ -38,18 +55,7 @@ if __name__ == '__main__':
     main()
 
     """
-    Changelog:
-        This should now work with 'should_work.ino'
-        I have not tried it yet, am afraid to break the code on the arduino as I don't have the code that is currently on it.
-
-        the mode is now handled by the driver and not the arduino
-
-    
-
     TODO:
-        Work on rocket flying   https://www.youtube.com/watch?v=ItN-K-WSCkM 
-        wellskate macro too
-
         try and transition all pyautogui keyboard funtions to custom_keyboard
         add a write function to custom_keyboard